requirements.txt
factordb-pycli
BigNumber
gmpy2
pycryptodome